+++ extensions/ohif-i18n/package.json
@@ -1,7 +1,7 @@
 {
   "name": "ohif-i18n",
   "version": "0.0.1",
-  "description": "OHIF extension internationalization",
+  "description": "OHIF extension for internationalization",
   "author": "OHIF",
   "license": "MIT",
   "repository": "OHIF/Viewers",
@@ -17,8 +17,8 @@
     "prepublishOnly": "npm run build"
   },
   "peerDependencies": {
-    "react-i18next": "^10.11.0",
-    "i18next": "^15.1.3"
+    "i18next": "^15.1.3",
+    "react-i18next": "^10.11.0"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
@@ -40,6 +40,7 @@
     "rollup": "^1.1.2",
     "rollup-plugin-babel": "^4.2.0",
     "rollup-plugin-commonjs": "^9.2.0",
+    "rollup-plugin-copy": "^2.0.1",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-resolve": "^4.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.0",
