--- extensions/ohif-i18n/src/locales/en-UK/Header.json
@@ -1,3 +1,0 @@
-{
-  "header_test": "Joao Grandao"
-}
