--- extensions/ohif-i18n/src/locales/en-UK/common.json
@@ -1,6 +1,0 @@
-{
-  "common": {
-    "Welcome to React": "Welcome to React and react-i18next",
-    "welcomeMsg": "welcome"
-  }
-}
