--- extensions/ohif-i18n/src/locales/en-US/Header.json
@@ -1,3 +1,0 @@
-{
-  "header_test": "$t(common:test)"
-}
