--- extensions/ohif-i18n/src/locales/en-US/common.json
@@ -1,3 +1,0 @@
-{
-  "tests___": "test common"
-}
