+++ extensions/ohif-i18n/src/locales/en/CineDialog.json
@@ -1,0 +1,18 @@
+{
+  "fps": "ingles fps",
+  "Skip to first image": "Skip to first image",
+  "Previous image": "Previous, image",
+  "Play / Stop": "Play, / Stop",
+  "Next image": "Next, image",
+  "Skip to last image": "Skip, to last image",
+  "Skip to first image": "Skip, to first image",
+  "Previous image": "Previous, image",
+  "Play / Stop": "Play, / Stop",
+  "Next image": "Next, image",
+  "Skip to last image": "Skip, to last image",
+  "Skip to first image": "Skip, to first image",
+  "Previous image": "Previous, image",
+  "Play / Stop": "Play, / Stop",
+  "Next image": "Next, image",
+  "Skip to last image": "Skip, to last image"
+}
