+++ extensions/ohif-i18n/src/locales/en/UK/Header.json
@@ -1,0 +1,3 @@
+{
+  "research_use": "INVESTIGATIONAL PURPOSES ONLY"
+}
