+++ extensions/ohif-i18n/src/locales/en/US/Header.json
@@ -1,0 +1,4 @@
+{
+  "options": "Erik Options",
+  "about": "About"
+}
