+++ extensions/ohif-i18n/src/locales/es/Buttons.json
@@ -36,5 +36,4 @@
   "Acquired": "Acquired",
   "Sagittal": "Sagittal",
   "Axial": "Axial"
-
 }
