+++ extensions/ohif-i18n/src/locales/es/CineDialog.json
@@ -1,0 +1,8 @@
+{
+  "fps": "fps",
+  "Skip to first image": "Avanza para la primera imagen",
+  "Previous image": "Imagen anterior",
+  "Play / Stop": "Play / Stop",
+  "Next image": "Próximo image",
+  "Skip to last image": "Pular para la ultima imagen"
+}
