+++ extensions/ohif-i18n/src/locales/es/common.json
