+++ extensions/ohif-vtk-extension/package.json
@@ -32,7 +32,8 @@
     "react-resize-detector": "^3.4.0",
     "redux": "^4.0.1",
     "i18next": "^15.1.3",
-    "react-i18next": "^10.11.0"
+    "react-i18next": "^10.11.0",
+    "react-viewerbase": "^0.6.1"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
@@ -54,7 +55,6 @@
     "prettier": "^1.15.3",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
-    "react-viewerbase": "^0.2.11",
     "rollup": "^1.1.2",
     "rollup-plugin-babel": "^4.2.0",
     "rollup-plugin-commonjs": "^9.2.0",
@@ -90,6 +90,7 @@
   "dependencies": {
     "@babel/runtime": "^7.2.0",
     "classnames": "^2.2.6",
+    "ohif-vtk-extension": "file:.yalc/ohif-vtk-extension",
     "react-vtkjs-viewport": "^0.0.7"
   }
 }
