+++ package.json
@@ -83,6 +83,7 @@
     "ohif-dicom-html-extension": "^0.0.2",
     "ohif-dicom-microscopy-extension": "^0.0.5",
     "ohif-dicom-pdf-extension": "^0.0.6",
+    "ohif-i18n": "file:.yalc/ohif-i18n",
     "ohif-vtk-extension": "^0.0.1",
     "oidc-client": "1.7.x",
     "prop-types": "^15.6.2",
