--- src/locales/en/CineDialog.json
@@ -1,18 +1,0 @@
-{
-  "fps":  "ingles fps",
-  "Skip to first image":  "Skip to first image",
-  "Previous image":  "Previous, image",
-  "Play / Stop":  "Play, / Stop",
-  "Next image":  "Next, image",
-  "Skip to last image":  "Skip, to last image",
-  "Skip to first image":  "Skip, to first image",
-  "Previous image":  "Previous, image",
-  "Play / Stop":  "Play, / Stop",
-  "Next image":  "Next, image",
-  "Skip to last image":  "Skip, to last image",
-  "Skip to first image":  "Skip, to first image",
-  "Previous image":  "Previous, image",
-  "Play / Stop":  "Play, / Stop",
-  "Next image":  "Next, image",
-  "Skip to last image":  "Skip, to last image"
-}
