--- src/locales/en/Header.json
@@ -1,8 +1,0 @@
-{
-  "preferences": "Preferences",
-  "options": "Options",
-  "about": "About",
-  "back_to_viewer": "Back to Viewer",
-  "study_list": "Study list",
-  "research_use": "INVESTIGATIONAL USE ONLY"
-}
