--- src/locales/en/UK/Header.json
@@ -1,3 +1,0 @@
-{
-  "header_research_use": "INVESTIGATIONAL PURPOSES ONLY"
-}
