--- src/locales/en/US/Header.json
@@ -1,4 +1,0 @@
-{
-  "options": "Options",
-  "about": "About"
-}
