--- src/locales/en/common.json
@@ -1,5 +1,0 @@
-{
-  "Welcome to React": "Welcome to React and react-i18next",
-  "welcomeMsg": "welcome",
-  "test": "Lunch time"
-}
