--- src/locales/es/CineDialog.json
@@ -1,8 +1,0 @@
-{
-  "fps":  "fps",
-  "Skip to first image":  "Avanza para la primera imagen",
-  "Previous image":  "Imagen anterior",
-  "Play / Stop":  "Play / Stop",
-  "Next image":  "Próximo image",
-  "Skip to last image":  "Pular para la ultima imagen"
-}
